Updating files edited
All files that I have edited are included. This should add a WIP button above the Inventory scroll area which, upon being clicked, adds a new button to the end of the list of items in the inventory scroll area and the order scroll area. This code cannot be tested in my work environment, but hopefully my edits perform as expected. I have also uploaded a basic itemButton object which can eventually replace the default button being added so that these buttons can be clicked to display information.